# server/docs/swagger_output_dev_new.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1/faculty/event/get-all-unapproved": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 3
                      },
                      "eventName": {
                        "type": "string",
                        "example": "WebVerse2.0"
                      },
                      "eventDescription": {
                        "type": "string",
                        "example": "fiuhgudasisnlgfdg"
                      },
                      "eventPoster": {
                        "type": "string",
                        "example": "url://123/img.jpg"
                      },
                      "eventDate": {
                        "type": "string",
                        "example": "10/10/2023"
                      },
                      "eventTime": {
                        "type": "string",
                        "example": "10:00"
                      },
                      "eventVenue": {
                        "type": "string",
                        "example": "MG Audi"
                      },
                      "eventOrganiser": {
                        "type": "string",
                        "example": "Android Club"
                      },
                      "participantCount": {
                        "type": "number",
                        "example": 12
                      },
                      "hostedBy": {
                        "type": "string",
                        "example": "Android Club"
                      },
                      "isApproved": {
                        "type": "boolean",
                        "example": false
                      },
                      "isRejected": {
                        "type": "boolean",
                        "example": false
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/faculty/event/get-all-approved": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "eventName": {
                        "type": "string",
                        "example": "WebVerse"
                      },
                      "eventDescription": {
                        "type": "string",
                        "example": "fiuhguisnlgfdg"
                      },
                      "eventPoster": {
                        "type": "string",
                        "example": "url://123/img.jpg"
                      },
                      "eventDate": {
                        "type": "string",
                        "example": "10/10/2023"
                      },
                      "eventTime": {
                        "type": "string",
                        "example": "10:00"
                      },
                      "eventVenue": {
                        "type": "string",
                        "example": "MG Audi"
                      },
                      "eventOrganiser": {
                        "type": "string",
                        "example": "Android Club"
                      },
                      "participantCount": {
                        "type": "number",
                        "example": 12
                      },
                      "hostedBy": {
                        "type": "string",
                        "example": "Android Club"
                      },
                      "isApproved": {
                        "type": "boolean",
                        "example": true
                      },
                      "isRejected": {
                        "type": "boolean",
                        "example": false
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/faculty/event/approve-event": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "eventId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 4
                    },
                    "eventName": {
                      "type": "string",
                      "example": "WebVerse3.0"
                    },
                    "eventDescription": {
                      "type": "string",
                      "example": "fiuhgudasisnlgfdg"
                    },
                    "eventPoster": {
                      "type": "string",
                      "example": "url://123/img.jpg"
                    },
                    "eventDate": {
                      "type": "string",
                      "example": "10/10/2023"
                    },
                    "eventTime": {
                      "type": "string",
                      "example": "10:00"
                    },
                    "eventVenue": {
                      "type": "string",
                      "example": "MG Audi"
                    },
                    "eventOrganiser": {
                      "type": "string",
                      "example": "Android Club"
                    },
                    "participantCount": {
                      "type": "number",
                      "example": 12
                    },
                    "hostedBy": {
                      "type": "string",
                      "example": "Android Club"
                    },
                    "isApproved": {
                      "type": "boolean",
                      "example": true
                    },
                    "isRejected": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/faculty/event/reject-event": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "eventId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 4
                    },
                    "eventName": {
                      "type": "string",
                      "example": "WebVerse3.0"
                    },
                    "eventDescription": {
                      "type": "string",
                      "example": "fiuhgudasisnlgfdg"
                    },
                    "eventPoster": {
                      "type": "string",
                      "example": "url://123/img.jpg"
                    },
                    "eventDate": {
                      "type": "string",
                      "example": "10/10/2023"
                    },
                    "eventTime": {
                      "type": "string",
                      "example": "10:00"
                    },
                    "eventVenue": {
                      "type": "string",
                      "example": "MG Audi"
                    },
                    "eventOrganiser": {
                      "type": "string",
                      "example": "Android Club"
                    },
                    "participantCount": {
                      "type": "number",
                      "example": 12
                    },
                    "hostedBy": {
                      "type": "string",
                      "example": "Android Club"
                    },
                    "isApproved": {
                      "type": "boolean",
                      "example": false
                    },
                    "isRejected": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}